Moved models and schema into db module for shorter import path
