+++ src/db/models.rs
@@ -1,6 +1,4 @@
-use super::schema::users;
-
-use diesel::sql_types::{Int4};
+use db::schema::users;
 
 #[derive(Queryable, Serialize, Deserialize, Insertable, Debug)]
 #[table_name="users"]
@@ -8,4 +6,3 @@ pub struct User {
     id: i32,
     email: String
 }
-
