+++ src/lib.rs
@@ -13,6 +13,4 @@ extern crate serde_json;
 
 pub mod config;
 pub mod db;
-pub mod models;
-pub mod schema;
 pub mod web;
