+++ src/web/app.rs
@@ -1,10 +1,8 @@
 use config::{AppConfig, VERSION};
 use diesel::pg::PgConnection;
-use diesel::prelude::*;
 use r2d2::Pool;
 use r2d2_diesel::ConnectionManager;
 use rocket::Rocket;
-use std;
 use web::members;
 
 pub fn mount_app(rocket: Rocket, config: AppConfig) -> Rocket {
@@ -13,7 +11,7 @@ pub fn mount_app(rocket: Rocket, config: AppConfig) -> Rocket {
     let pool = Pool::new(manager).expect("Failed to create pool.");
     rocket
         .mount("/", routes![version])
-        .mount("/member", members::routes())
+        .mount("/members", members::routes())
         .manage(pool)
 }
